blueprints + added pytest cache to git ignore
