+++ .gitignore
@@ -2,3 +2,4 @@
 __pycache__
 .DS_Store
 *.db
+.pytest_cache
